refactor(frontend): clarify orchestrator file names
