+++ packages/frontend/src/state/operations/agent-orchestrator/support/orchestrator-ports.ts
@@ -4,8 +4,7 @@ import type {
   TaskWorktreeSummary,
 } from "@openducktor/contracts";
 import type { QueryClient } from "@tanstack/react-query";
-import { appQueryClient } from "@/lib/query-client";
-import { host } from "../../shared/host";
+import type { host } from "../../shared/host";
 
 export type AgentOrchestratorHostPort = {
   agentSessionUpsert: (
@@ -27,19 +26,3 @@ export type AgentOrchestratorDependencies = {
   hostPort: AgentOrchestratorHostPort;
   runtimeHostPort: AgentOrchestratorRuntimeHostPort;
 };
-
-export const createDefaultAgentOrchestratorDependencies = (): AgentOrchestratorDependencies => ({
-  queryClient: appQueryClient,
-  hostPort: {
-    agentSessionUpsert: (repoPath, taskId, record) =>
-      host.agentSessionUpsert(repoPath, taskId, record),
-    agentSessionStop: async (target) => {
-      await host.agentSessionStop(target);
-    },
-    taskWorktreeGet: (repoPath, taskId) => host.taskWorktreeGet(repoPath, taskId),
-  },
-  runtimeHostPort: {
-    buildStart: (...args) => host.buildStart(...args),
-    runtimeEnsure: (...args) => host.runtimeEnsure(...args),
-  },
-});
